refactor: AiAreaServiceを責務分割 (#385)
- AiArea::SpotFinder: 円内スポット検索を担当
- AiArea::PromptBuilder: プロンプト生成を担当
- AiAreaService: オーケストレーションに専念

337行 → 169行 + 96行 + 92行（各クラス単一責務）
